Alex: Rename class; throw exceptions for invalid inputs
