+++ Intersect.Server/Core/ConsoleService.ConsoleThread.cs
@@ -51,7 +51,7 @@ namespace Intersect.Server.Core
                 Parser.Register<PowerCommand>();
                 Parser.Register<UnbanCommand>();
                 Parser.Register<UnmuteCommand>();
-                Parser.Register<SetVarCommand>();
+                Parser.Register<SetVariableCommand>();
             }
 
             public CommandParser Parser { get; }
